move specs to specs
