+++ corehq/apps/userreports/reports/factory.py
@@ -1,15 +1,10 @@
 import json
-from jsonobject import JsonObject, StringProperty
 from jsonobject.exceptions import BadValueError
-from sqlagg import SumColumn
-from sqlagg.columns import SimpleColumn
-from corehq.apps.reports.sqlreport import DatabaseColumn
 from corehq.apps.reports_core.filters import DatespanFilter, ChoiceListFilter, Choice
 from corehq.apps.userreports.exceptions import BadSpecError
-from corehq.apps.userreports.reports.filters import DateFilterValue, ChoiceListFilterValue
 from django.utils.translation import ugettext as _
 from corehq.apps.userreports.reports.specs import FilterSpec, ChoiceListFilterSpec, PieChartSpec, \
-    MultibarAggregateChartSpec, ChartSpec, MultibarChartSpec
+    MultibarAggregateChartSpec, MultibarChartSpec, ReportFilter, ReportColumn
 
 
 def _build_date_filter(spec):
@@ -93,33 +88,3 @@ class ChartFactory(object):
                 json.dumps(spec, indent=2),
                 str(e),
             ))
-
-
-class ReportFilter(JsonObject):
-    type = StringProperty(required=True)
-    slug = StringProperty(required=True)
-    field = StringProperty(required=True)
-    display = StringProperty()
-
-    def create_filter_value(self, value):
-        return {
-            'date': DateFilterValue,
-            'choice_list': ChoiceListFilterValue,
-        }[self.type](self, value)
-
-
-class ReportColumn(JsonObject):
-    type = StringProperty(required=True)
-    display = StringProperty()
-    field = StringProperty(required=True)
-    aggregation = StringProperty(required=True)
-
-    def get_sql_column(self):
-        # todo: find a better home for this
-        sqlagg_column_map = {
-            'sum': SumColumn,
-            'simple': SimpleColumn,
-        }
-        return DatabaseColumn(self.display, sqlagg_column_map[self.aggregation](self.field),
-                              sortable=False, data_slug=self.field)
-
